ver 2.08: incorrect mask applied at loading saved presets implied wrong channel settings
Thanks to szkudlik for finding this issues. 
Settings are applied BEFORE the source should not be applied to new channels. The app is design to change the order of the channels only, not changing the ways that the sources are physically connected, it should never change setting of any preamp.